requirements.txt
numpy>=1.23
pandas>=2.0
yfinance>=0.2
scikit-learn>=1.2
matplotlib>=3.7
torch>=2.0
tqdm>=4.66
pyyaml>=6.0
